+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [14.3.0] - 2021-03-24
 ### Added
 - AS3 - Remove trait which is outside class
 - PDF vector export